backend/requirements.txt
# Core FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-jose[cryptography]==3.3.0

# AI/ML Models
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2
faiss-cpu==1.7.4
spacy==3.7.2

# Document Processing
pymupdf==1.23.21
pytesseract==0.3.10
Pillow==10.2.0
opencv-python==4.9.0.80
python-docx==1.1.0

# NLP & Translation
langdetect==1.0.9
sentencepiece==0.1.99
sacremoses==0.1.1

# Data Processing
numpy==1.26.3
pandas==2.2.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-magic-bin==0.4.14
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Logging & Monitoring
python-json-logger==2.0.7
